Update runtimeIcon to use quarto codicon
