+++ src/vs/workbench/contrib/positronConsole/common/sessionDisplayUtils.ts
@@ -1,0 +1,22 @@
+/*---------------------------------------------------------------------------------------------
+ *  Copyright (C) 2026 Posit Software, PBC. All rights reserved.
+ *  Licensed under the Elastic License 2.0. See LICENSE.txt for license information.
+ *--------------------------------------------------------------------------------------------*/
+
+import { URI } from '../../../../base/common/uri.js';
+import { IModelService } from '../../../../editor/common/services/model.js';
+import { isQuartoDocument } from '../../positronQuarto/common/positronQuartoConfig.js';
+
+/**
+ * Checks if a notebook URI belongs to a Quarto/RMarkdown document by looking
+ * up the editor model's language ID. This works for both saved files (where the
+ * URI path has .qmd/.rmd) and untitled files (where the model's language ID is
+ * set to "quarto" or "rmd" by the Quarto extension).
+ */
+export function isQuartoSession(notebookUri: URI | undefined, modelService: IModelService): boolean {
+	if (!notebookUri) {
+		return false;
+	}
+	const model = modelService.getModel(notebookUri);
+	return isQuartoDocument(notebookUri.path, model?.getLanguageId());
+}
